requirements.txt
Pillow==9.5.0
scikit-learn==1.2.2
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python-headless==4.8.0.74
pyautogui==0.9.53
